use janome instead of Mecab.
