+++ summpy/tools.py
@@ -6,8 +6,6 @@ import re
 import json
 import subprocess
 
-import MeCab
-
 
 def tree_encode(obj, encoding='utf-8'):
     type_ = type(obj)
@@ -30,76 +28,6 @@ def ppj(obj, sort_keys=False):
         tree_encode(obj), indent=2, ensure_ascii=False, sort_keys=sort_keys
     )
     print obj_str
-
-
-# ここからmecab関連
-
-_mecab = MeCab.Tagger()
-# 品詞,品詞細分類1,品詞細分類2,品詞細分類3,活用形,活用型,原形,読み,発音
-_mecab_feat_labels = 'pos cat1 cat2 cat3 conj conj_t orig read pron'.split(' ')
-
-
-def _mecab_parse_feat(feat):
-    return dict(zip(_mecab_feat_labels, feat.split(',')))
-
-
-def _mecab_node2seq(node, decode_surface=True, feat_dict=True,
-                    mecab_encoding='utf-8'):
-    # MeCab.Nodeはattributeを変更できない。
-    while node:
-        if decode_surface:
-            node._surface = node.surface.decode(mecab_encoding)
-        if feat_dict:  # 品詞の情報をdictで保存
-            node.feat_dict = _mecab_parse_feat(
-                node.feature.decode(mecab_encoding)
-            )
-        yield node
-        node = node.next
-
-
-def is_stopword(n):  # <- mecab node
-    if len(n._surface) == 0:
-        return True
-    elif re.search(ur'^[\s!-@\[-`\{-~　、-〜！-＠［-｀]+$', n._surface):
-        return True
-    elif re.search(ur'^(接尾|非自立)', n.feat_dict['cat1']):
-        return True
-    elif u'サ変・スル' == n.feat_dict['conj'] or u'ある' == n.feat_dict['orig']:
-        return True
-    elif re.search(ur'^(名詞|動詞|形容詞)', n.feat_dict['pos']):
-        return False
-    else:
-        return True
-
-
-def not_stopword(n):  # <- mecab node
-    return not is_stopword(n)
-
-
-def _node2word(n):  # <- mecab node
-    return n._surface
-
-
-def _node2norm_word(n):  # mecab node
-    if n.feat_dict['orig'] != '*':
-        return n.feat_dict['orig']
-    else:
-        return n._surface
-
-
-def word_segmenter_ja(sent, node_filter=not_stopword,
-                      node2word=_node2norm_word, mecab_encoding='utf-8'):
-    if type(sent) == unicode:
-        sent = sent.encode(mecab_encoding)
-
-    nodes = list(
-        _mecab_node2seq(_mecab.parseToNode(sent), mecab_encoding=mecab_encoding)
-    )
-    if node_filter:
-        nodes = [n for n in nodes if node_filter(n)]
-    words = [node2word(n) for n in nodes]
-
-    return words
 
 
 def fix_paren_sents(sents):
@@ -155,14 +83,6 @@ def sent_splitter_ja(text, fix_paren=True):  # type(text) == unicode
     return sents
 
 
-def test_mecab():
-    text = u'今日はいい天気ですね。太郎は今日学校に行こうとしています。'
-    sents = sent_splitter_ja(text)
-    for s in sents:
-        ws = word_segmenter_ja(s)
-        print '|'.join(ws)
-
-
 def l2norm(xs):
     return math.sqrt(sum(x * x for x in xs))
 
@@ -175,7 +95,3 @@ def cos_sim(v1, v2):
         a += v1[k] * (v2[k] if k in v2 else 0)
     b = l2norm(v1.values()) * l2norm(v2.values())
     return a / b
-
-
-if __name__ == '__main__':
-    test_mecab()
